Mongrel monkey patch for Rails 2.3.8-2.3.11
When deploying with mongrel, Rails 2.3.8 will run in a number of
problems, described and discussed here:

   https://rails.lighthouseapp.com/projects/8994/tickets/4690

A first solution was suggested here:

  https://gist.github.com/471663

An enhanced solution was suggested here:

  https://gist.github.com/826692

This commit implements the latter.

This is actual especially if you are deploying FFC onto shared hosting
environments where you have no direct control of the entire software
stack. For instance, A2Hosting does not support phusion, but offers
Mongrel, and therefore this patch is necessary.